requirements.txt
Flask==3.0.3
requests==2.31.0
python-dotenv==1.0.1
gunicorn==21.2.0
pytest==8.2.0
pytest-flask==1.3.0
gTTS==2.5.1
black>=24.0.0
pylint>=3.0.0
google-cloud-logging==3.10.0
google-cloud-storage==2.16.0
google-cloud-secret-manager==2.20.0
google-cloud-error-reporting==1.11.1
google-generativeai>=0.8.0
